refactor(frontend): reset form onSuccess
